commonroad_environment/install/dataset-converters/requirements.txt
numpy>=1.18.2
commonroad-io>=2020.3
pandas>=0.24.2
scipy>=1.4.1
ruamel.yaml>=0.16.10